Initial destroy method
